+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Added `ResidentTexture`, `TextureHandle` and `Texture::resident_if_supported()` for bindless textures.
  - Buffers no longer require their content to be `Send` or `'static` (except for types with the `Any` suffix).
  - Renamed `texture::TextureType` to `Dimensions`.
+ - Added `Context::get_opengl_version()` and deprecated `get_version()`.
 
 ## Version 0.6.6 (2015-07-03)